+++ libs/blocks/src/lib/stat/index.tsx
@@ -1,10 +1,12 @@
 import { ReactNode } from 'react';
-import NumbersWithTitle from './numbers-with-title';
-import NumbersOnly from './numbers-only';
+import V1StatBlock from './v1';
+import V2StatBlock from './v2';
+import V3StatBlock from './v3';
 
 export type StatBlockVariants = {
-  NumbersWithTitle: typeof NumbersWithTitle;
-  NumbersOnly: typeof NumbersOnly;
+  V1: typeof V1StatBlock;
+  V2: typeof V2StatBlock;
+  V3: typeof V3StatBlock;
 };
 
 export const StatBlock: StatBlockVariants = ({
@@ -15,7 +17,8 @@ export const StatBlock: StatBlockVariants = ({
   return children;
 };
 
-StatBlock.NumbersWithTitle = NumbersWithTitle;
-StatBlock.NumbersOnly = NumbersOnly;
+StatBlock.V1 = V1StatBlock;
+StatBlock.V2 = V2StatBlock;
+StatBlock.V3 = V3StatBlock;
 
 export default StatBlock;